pmm: isolate the bitmap code
